Add files via upload
A scatter plot of the per capita spending per student vs the average math score. An overlay of a linear regression is shown in the black line; we can see a positive correlation between per capita spending and math scores. There are some extreme points, which indicates removal of outliers might be necessary before moving into a machine learning phase.